Reorganize the order of items in the Rust code for consistency
