+++ crates/huub/src/helpers/linear_transform.rs
@@ -2,9 +2,7 @@
 
 use std::ops::{Add, Mul, Neg};
 
-use rangelist::RangeList;
-
-use crate::{helpers::div_ceil, IntVal, LitMeaning, NonZeroIntVal};
+use crate::{helpers::div_ceil, IntSetVal, IntVal, LitMeaning, NonZeroIntVal};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 /// An integer linear transformation of a discrete value.
@@ -23,10 +21,6 @@ impl LinearTransform {
 	pub fn new(scale: NonZeroIntVal, offset: IntVal) -> Self {
 		Self { scale, offset }
 	}
-	/// Creates a new linear transformation with the given scale and no offset.
-	pub fn scaled(scale: NonZeroIntVal) -> Self {
-		Self { scale, offset: 0 }
-	}
 	/// Creates a new linear transformation with the given offset and no scale.
 	pub fn offset(offset: IntVal) -> Self {
 		Self {
@@ -40,31 +34,9 @@ impl LinearTransform {
 		self.scale.get() > 0
 	}
 
-	/// Perform the linear transformation on a value.
-	pub fn transform(&self, val: IntVal) -> IntVal {
-		(val * self.scale.get()) + self.offset
-	}
-
-	/// Perform the linear tranformation on a `LitMeaning`.
-	pub fn transform_lit(&self, mut lit: LitMeaning) -> LitMeaning {
-		let mut transformer = *self;
-		if !self.positive_scale() {
-			// Make positive by doing `*-1` on lit meaning and transformer
-			(lit, transformer) = match lit {
-				// -x >= i === x <= -i === x < -i + 1
-				LitMeaning::GreaterEq(i) => (LitMeaning::Less(-i + 1), -transformer),
-				// -x < i === x > -i === x >= -i + 1
-				LitMeaning::Less(i) => (LitMeaning::GreaterEq(-i + 1), -transformer),
-				_ => (lit, transformer),
-			};
-		}
-
-		match lit {
-			LitMeaning::Eq(v) => LitMeaning::Eq(transformer.transform(v)),
-			LitMeaning::NotEq(v) => LitMeaning::NotEq(transformer.transform(v)),
-			LitMeaning::GreaterEq(v) => LitMeaning::GreaterEq(transformer.transform(v)),
-			LitMeaning::Less(v) => LitMeaning::Less(transformer.transform(v)),
-		}
+	/// Returns whether a value remains an integer after reversing the transformation.
+	pub fn rev_remains_integer(&self, val: IntVal) -> bool {
+		(val - self.offset) % self.scale.get() == 0
 	}
 
 	/// Perform the reverse linear transformation on a value.
@@ -72,9 +44,27 @@ impl LinearTransform {
 		(val - self.offset) / self.scale.get()
 	}
 
-	/// Returns whether a value remains an integer after reversing the transformation.
-	pub fn rev_remains_integer(&self, val: IntVal) -> bool {
-		(val - self.offset) % self.scale.get() == 0
+	/// Reverse the linear transformation on a set of integer values.
+	pub(crate) fn rev_transform_int_set(&self, mask: &IntSetVal) -> IntSetVal {
+		let get_val = |meaning| match meaning {
+			LitMeaning::GreaterEq(i) => i,
+			LitMeaning::Less(i) => i - 1,
+			_ => unreachable!(),
+		};
+
+		mask.iter()
+			.map(|r| {
+				let a = get_val(
+					self.rev_transform_lit(LitMeaning::GreaterEq(*r.start()))
+						.unwrap(),
+				);
+				let b = get_val(
+					self.rev_transform_lit(LitMeaning::Less(r.end() + 1))
+						.unwrap(),
+				);
+				a.min(b)..=a.max(b)
+			})
+			.collect()
 	}
 
 	/// Perform the reverse linear tranformation for a `LitMeaning`.
@@ -122,46 +112,35 @@ impl LinearTransform {
 			))),
 		}
 	}
-
-	/// Reverse the linear transformation on a set of integer values.
-	pub(crate) fn rev_transform_mask(&self, mask: &RangeList<IntVal>) -> RangeList<IntVal> {
-		let get_val = |meaning| match meaning {
-			LitMeaning::GreaterEq(i) => i,
-			LitMeaning::Less(i) => i - 1,
-			_ => unreachable!(),
-		};
-
-		mask.iter()
-			.map(|r| {
-				let a = get_val(
-					self.rev_transform_lit(LitMeaning::GreaterEq(*r.start()))
-						.unwrap(),
-				);
-				let b = get_val(
-					self.rev_transform_lit(LitMeaning::Less(r.end() + 1))
-						.unwrap(),
-				);
-				a.min(b)..=a.max(b)
-			})
-			.collect()
+	/// Creates a new linear transformation with the given scale and no offset.
+	pub fn scaled(scale: NonZeroIntVal) -> Self {
+		Self { scale, offset: 0 }
 	}
-}
 
-impl Default for LinearTransform {
-	fn default() -> Self {
-		Self {
-			scale: NonZeroIntVal::new(1).unwrap(),
-			offset: 0,
+	/// Perform the linear transformation on a value.
+	pub fn transform(&self, val: IntVal) -> IntVal {
+		(val * self.scale.get()) + self.offset
+	}
+
+	/// Perform the linear tranformation on a `LitMeaning`.
+	pub fn transform_lit(&self, mut lit: LitMeaning) -> LitMeaning {
+		let mut transformer = *self;
+		if !self.positive_scale() {
+			// Make positive by doing `*-1` on lit meaning and transformer
+			(lit, transformer) = match lit {
+				// -x >= i === x <= -i === x < -i + 1
+				LitMeaning::GreaterEq(i) => (LitMeaning::Less(-i + 1), -transformer),
+				// -x < i === x > -i === x >= -i + 1
+				LitMeaning::Less(i) => (LitMeaning::GreaterEq(-i + 1), -transformer),
+				_ => (lit, transformer),
+			};
 		}
-	}
-}
 
-impl Neg for LinearTransform {
-	type Output = Self;
-	fn neg(self) -> Self::Output {
-		Self {
-			scale: NonZeroIntVal::new(-self.scale.get()).unwrap(),
-			offset: -self.offset,
+		match lit {
+			LitMeaning::Eq(v) => LitMeaning::Eq(transformer.transform(v)),
+			LitMeaning::NotEq(v) => LitMeaning::NotEq(transformer.transform(v)),
+			LitMeaning::GreaterEq(v) => LitMeaning::GreaterEq(transformer.transform(v)),
+			LitMeaning::Less(v) => LitMeaning::Less(transformer.transform(v)),
 		}
 	}
 }
@@ -177,6 +156,15 @@ impl Add<IntVal> for LinearTransform {
 	}
 }
 
+impl Default for LinearTransform {
+	fn default() -> Self {
+		Self {
+			scale: NonZeroIntVal::new(1).unwrap(),
+			offset: 0,
+		}
+	}
+}
+
 impl Mul<NonZeroIntVal> for LinearTransform {
 	type Output = Self;
 
@@ -184,6 +172,17 @@ impl Mul<NonZeroIntVal> for LinearTransform {
 		LinearTransform {
 			scale: NonZeroIntVal::new(self.scale.get() * rhs.get()).unwrap(),
 			offset: self.offset * rhs.get(),
+		}
+	}
+}
+
+impl Neg for LinearTransform {
+	type Output = Self;
+
+	fn neg(self) -> Self::Output {
+		Self {
+			scale: NonZeroIntVal::new(-self.scale.get()).unwrap(),
+			offset: -self.offset,
 		}
 	}
 }
