+++ crates/huub/src/helpers/opt_field.rs
@@ -16,6 +16,13 @@ pub(crate) struct OptField<const B: usize, T> {
 	value: [T; B],
 }
 
+impl<T> OptField<1, T> {
+	/// Creates a new `OptField` with the given value.
+	pub(crate) fn new(value: T) -> Self {
+		Self { value: [value] }
+	}
+}
+
 impl<const B: usize, T> OptField<B, T> {
 	#[inline]
 	/// Return the value of the `OptField`, if it exists.
@@ -24,10 +31,17 @@ impl<const B: usize, T> OptField<B, T> {
 	}
 }
 
-impl<T> OptField<1, T> {
-	/// Creates a new `OptField` with the given value.
-	pub(crate) fn new(value: T) -> Self {
-		Self { value: [value] }
+impl<const B: usize, T: Clone> Clone for OptField<B, T> {
+	fn clone(&self) -> Self {
+		Self {
+			value: self.value.clone(),
+		}
+	}
+}
+
+impl<T> Default for OptField<0, T> {
+	fn default() -> Self {
+		Self { value: [] }
 	}
 }
 
@@ -39,30 +53,16 @@ impl<T: Default> Default for OptField<1, T> {
 	}
 }
 
-impl<T> Default for OptField<0, T> {
-	fn default() -> Self {
-		Self { value: [] }
+impl<const B: usize, T: Eq> Eq for OptField<B, T> {}
+
+impl<const B: usize, T: Hash> Hash for OptField<B, T> {
+	fn hash<H: Hasher>(&self, state: &mut H) {
+		self.value.iter().for_each(|v| v.hash(state));
 	}
 }
 
 impl<const B: usize, T: PartialEq> PartialEq for OptField<B, T> {
 	fn eq(&self, other: &Self) -> bool {
 		self.value == other.value
-	}
-}
-
-impl<const B: usize, T: Eq> Eq for OptField<B, T> {}
-
-impl<const B: usize, T: Clone> Clone for OptField<B, T> {
-	fn clone(&self) -> Self {
-		Self {
-			value: self.value.clone(),
-		}
-	}
-}
-
-impl<const B: usize, T: Hash> Hash for OptField<B, T> {
-	fn hash<H: Hasher>(&self, state: &mut H) {
-		self.value.iter().for_each(|v| v.hash(state));
 	}
 }
