+++ crates/huub/src/solver/engine/bool_to_int.rs
@@ -19,26 +19,6 @@ pub(crate) struct BoolToIntMap {
 }
 
 impl BoolToIntMap {
-	/// Insert a range of Boolean variables to map them to the integer variable
-	/// for which they are eagerly created to represent conditions for.
-	pub(crate) fn insert_eager(&mut self, range: VarRange, var: IntVarRef) {
-		if range.is_empty() {
-			return;
-		}
-		if self.eager.is_empty() || self.eager.last().unwrap().0.end() < range.start() {
-			self.eager.push((range, var));
-			return;
-		}
-		panic!("Literal Mapping not added in the correct order")
-	}
-
-	/// Insert a mapping of a lazily created Boolean variable to the integer
-	/// variable and the meaning of the literal on the integer variable.
-	pub(crate) fn insert_lazy(&mut self, var: RawVar, iv: IntVarRef, lit: LitMeaning) {
-		let x = self.lazy.insert(var, (iv, lit));
-		debug_assert_eq!(x, None, "lazy literal already exists");
-	}
-
 	/// Return the integer variable the given Boolean variable represents a
 	/// condition for, if any. If the Boolean variable was lazily created, then
 	/// also return the [`LitMeaning`] of the literal.
@@ -66,5 +46,24 @@ impl BoolToIntMap {
 				.get(&var)
 				.map(|(int_var, meaning)| (*int_var, Some(meaning.clone())))
 		}
+	}
+	/// Insert a range of Boolean variables to map them to the integer variable
+	/// for which they are eagerly created to represent conditions for.
+	pub(crate) fn insert_eager(&mut self, range: VarRange, var: IntVarRef) {
+		if range.is_empty() {
+			return;
+		}
+		if self.eager.is_empty() || self.eager.last().unwrap().0.end() < range.start() {
+			self.eager.push((range, var));
+			return;
+		}
+		panic!("Literal Mapping not added in the correct order")
+	}
+
+	/// Insert a mapping of a lazily created Boolean variable to the integer
+	/// variable and the meaning of the literal on the integer variable.
+	pub(crate) fn insert_lazy(&mut self, var: RawVar, iv: IntVarRef, lit: LitMeaning) {
+		let x = self.lazy.insert(var, (iv, lit));
+		debug_assert_eq!(x, None, "lazy literal already exists");
 	}
 }
