+++ crates/huub/src/solver/engine/int_var.rs
@@ -20,8 +20,29 @@ use crate::{
 		view::{BoolViewInner, IntViewInner},
 		IntView,
 	},
-	BoolView, Clause, IntVal, LinearTransform, NonZeroIntVal, Solver,
+	BoolView, Clause, IntSetVal, IntVal, LinearTransform, NonZeroIntVal, Solver,
 };
+
+/// An entry in the [`DirectStorage`] that can be used to access the
+/// representation of an equality condition, or insert a new literal to
+/// represent the condition otherwise.
+enum DirectEntry<'a> {
+	/// The condition is already stored in the [`DirectStorage`].
+	Occupied(BoolViewInner),
+	/// The condition is not yet stored in the [`DirectStorage`].
+	Vacant(VacantEntry<'a, IntVal, RawVar>),
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+/// The structure that stores the equality conditions. Equality conditions can
+/// either be eagerly crated, and stored as a range of variables, or lazily
+/// created and stored in a [`HashMap`] once created.
+pub(crate) enum DirectStorage {
+	/// Variables for all equality conditions are eagerly created and stored in order
+	Eager(VarRange),
+	/// Variables for equality conditions are lazily created and stored in a hashmap
+	Lazy(HashMap<IntVal, RawVar>),
+}
 
 /// A type to represent when certain literals are created
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -30,11 +51,6 @@ pub(crate) enum EncodingType {
 	Eager,
 	/// The literal is created the first time it is mentioned
 	Lazy,
-}
-
-index_vec::define_index_type! {
-	/// Identifies an integer variable in a [`Solver`]
-	pub struct IntVarRef = u32;
 }
 
 #[derive(Debug, PartialEq, Eq, Clone)]
@@ -60,206 +76,249 @@ pub(crate) struct IntVar {
 	upper_bound: TrailedInt,
 }
 
-impl IntVar {
-	/// Create a new integer variable within the given solver, which the given
-	/// domain. The `order_encoding` and `direct_encoding` parameters determine
-	/// whether literals to reason about the integer variables are created eagerly
-	/// or lazily.
-	pub(crate) fn new_in<Oracle: PropagatingSolver<Engine>>(
-		slv: &mut Solver<Oracle>,
-		domain: RangeList<IntVal>,
-		order_encoding: EncodingType,
-		direct_encoding: EncodingType,
-	) -> IntView {
-		let orig_domain_len: usize = domain
-			.iter()
-			.map(|x| (x.end() - x.start() + 1) as usize)
-			.sum();
-		assert!(
-			orig_domain_len != 0,
-			"Unable to create integer variable empty domain"
-		);
-		if orig_domain_len == 1 {
-			return IntView(IntViewInner::Const(*domain.lower_bound().unwrap()));
-		}
-		let lb = *domain.lower_bound().unwrap();
-		let ub = *domain.upper_bound().unwrap();
-		if orig_domain_len == 2 {
-			let lit = slv.oracle.new_lit();
-			return IntView(IntViewInner::Bool {
-				transformer: LinearTransform {
-					scale: NonZeroIntVal::new(ub - lb).unwrap(),
-					offset: lb,
-				},
-				lit,
-			});
-		}
-		debug_assert!(
-			direct_encoding != EncodingType::Eager || order_encoding == EncodingType::Eager
-		);
+/// The definition given to a lazily created literal.
+pub(crate) struct LazyLitDef {
+	/// The meaning that the literal is meant to represent.
+	pub(crate) meaning: LitMeaning,
+	/// The variable that represent:
+	/// - if `meaning` is `LitMeaning::Less(j)`, then `prev` contains the literal
+	///   `< i` where `i` is the value right before `j` in the storage.
+	/// - if `meaning` is `LitMeaning::Eq(k)`, then `prev` contains the literal
+	///   `<j`.
+	pub(crate) prev: Option<RawVar>,
+	/// The variable that represent the literal `< k` where `k` is the value right
+	/// after the value represented by the literal.
+	pub(crate) next: Option<RawVar>,
+}
 
-		let upper_bound = slv.engine_mut().state.trail.track_int(ub);
-		let order_encoding = match order_encoding {
-			EncodingType::Eager => OrderStorage::Eager {
-				lower_bound: slv.engine_mut().state.trail.track_int(lb),
-				storage: slv.oracle.new_var_range(orig_domain_len - 1),
+#[derive(Clone, Debug, PartialEq, Eq)]
+/// A storage structure to manage lazily created order literals for an integer
+/// variable.
+pub(crate) struct LazyOrderStorage {
+	/// The index of the node with the minimum value in the storage.
+	min_index: u32,
+	/// The index of the node with the maximum value in the storage.
+	max_index: u32,
+	/// The index of the node that currently represents the lower bound of the
+	/// integer variable.
+	lb_index: TrailedInt,
+	/// The index of the node that currently represents the upper bound of the
+	/// integer variable.
+	ub_index: TrailedInt,
+	/// The storage of all currently created nodes containing the order literals
+	/// for the integer variable.
+	storage: Vec<OrderNode>,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+/// The meaning of a literal in the context of a [`IntVar`] `x`.
+pub enum LitMeaning {
+	/// Literal representing the condition `x = i`.
+	Eq(IntVal),
+	/// Literal representing the condition `x ≠ i`.
+	NotEq(IntVal),
+	/// Literal representing the condition `x ≥ i`.
+	GreaterEq(IntVal),
+	/// Literal representing the condition `x < i`.
+	Less(IntVal),
+}
+
+#[derive(Debug)]
+/// An entry in [`OrderStorage`] that can be used to access the representation
+/// of an inequality condition, or insert a new literal to represent the
+/// condition otherwise.
+enum OrderEntry<'a> {
+	/// Entry already exists and was eagerly created.
+	Eager(&'a VarRange, usize),
+	/// Entry already exists and was lazily created.
+	Occupied {
+		/// Reference to the storage where the entry is stored.
+		storage: &'a mut LazyOrderStorage,
+		/// The index of the node in the storage that the entry points to.
+		index: u32,
+		/// An iterator pointing at the range in the domain in which the value of
+		/// which the value of the entry is part.
+		range_iter: RangeIter<'a>,
+	},
+	/// Entry does not exist and can be lazily created.
+	Vacant {
+		/// Reference to the storage where the new entry will be created.
+		storage: &'a mut LazyOrderStorage,
+		/// The index of the node that contains the value right before the new entry
+		/// that will be created.
+		prev_index: IntVal,
+		/// An iterator pointing at the range in the domain in which the value of
+		/// which the value of the new entry is part.
+		range_iter: RangeIter<'a>,
+		/// The value for which the entry will be created.
+		val: IntVal,
+	},
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+/// Type useed to store individual entries in [`LazyOrderStorage`].
+pub(crate) struct OrderNode {
+	/// The value for which `var` represents `x < val`.
+	val: IntVal,
+	/// The variable representing `x < val`.
+	var: RawVar,
+	/// Whether there is a node with a value less than `val`.
+	has_prev: bool,
+	/// The index of the node with a value less than `val`.
+	prev: u32,
+	/// Whether there is a node with a value greater than `val`.
+	has_next: bool,
+	/// The index of the node with a value greater than `val`.
+	next: u32,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+#[allow(
+	variant_size_differences,
+	reason = "TODO: Investigate if using Box improves performance"
+)]
+/// The storage used to store the variables for the inequality conditions.
+pub(crate) enum OrderStorage {
+	/// Variables for all inequality conditions are eagerly created and stored in
+	/// order.
+	Eager {
+		/// A trailed integer that represents the currently lower bound of the
+		/// variable.
+		lower_bound: TrailedInt,
+		/// The range of Boolean variables that represent the inequality conditions.
+		storage: VarRange,
+	},
+	/// Variables for inequality conditions are lazily created and specialized
+	/// noded structure, a [`LazyOrderStorage`].
+	Lazy(LazyOrderStorage),
+}
+
+/// Type alias for an iterator that yields the ranges of a [`RangeList`], which
+/// is used to represent the domains of an integer variable.
+type RangeIter<'a> = Peekable<
+	Map<
+		<&'a RangeList<IntVal> as IntoIterator>::IntoIter,
+		fn(RangeInclusive<&'a IntVal>) -> RangeInclusive<IntVal>,
+	>,
+>;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// A direction to search in.
+enum SearchDirection {
+	/// Search from low to high.
+	Increasing,
+	/// Search from high to low.
+	Decreasing,
+}
+
+impl DirectEntry<'_> {
+	/// Extract the [`BoolViewInner`] if the entry is occupied, or insert a new
+	/// variable using the given function.
+	fn or_insert_with(self, f: impl FnOnce() -> RawVar) -> BoolViewInner {
+		match self {
+			DirectEntry::Occupied(bv) => bv,
+			DirectEntry::Vacant(no_entry) => {
+				let v = f();
+				let _ = no_entry.insert(v);
+				BoolViewInner::Lit(v.into())
+			}
+		}
+	}
+}
+
+impl DirectStorage {
+	/// Locate the position in the [`DirectStorage`] that would be used to store
+	/// the representation of the condition `= i`. The method will return a
+	/// [`DirectEntry`] object that can be used to access the condition as a
+	/// [`BoolViewInner`] if it already exists, or insert a new literal to
+	/// represent the condition otherwise.
+	///
+	/// The given `domain` is (in the case of eager creation) used to determine
+	/// the offset of the variable in the `VarRange`.
+	fn entry(&mut self, domain: &RangeList<IntVal>, i: IntVal) -> DirectEntry<'_> {
+		match self {
+			DirectStorage::Eager(vars) => {
+				// Calculate the offset in the VarRange
+				let mut offset = Some(-1); // -1 to account for the lower bound
+				for r in domain.iter() {
+					if i < *r.start() {
+						offset = None;
+						break;
+					} else if r.contains(&i) {
+						offset = Some(offset.unwrap() + i - r.start());
+						break;
+					} else {
+						offset = Some(offset.unwrap() + r.end() - r.start() + 1);
+					}
+				}
+				if let Some(offset) = offset {
+					debug_assert!(
+						(offset as usize) < vars.len(),
+						"var range offset, {}, must be in [{}, {})",
+						offset,
+						0,
+						vars.len(),
+					);
+					DirectEntry::Occupied(BoolViewInner::Lit(vars.index(offset as usize).into()))
+				} else {
+					DirectEntry::Occupied(BoolViewInner::Const(false))
+				}
+			}
+			DirectStorage::Lazy(map) => match map.entry(i) {
+				hash_map::Entry::Occupied(entry) => {
+					DirectEntry::Occupied(BoolViewInner::Lit((*entry.get()).into()))
+				}
+				hash_map::Entry::Vacant(no_entry) => {
+					if domain.contains(&i) {
+						DirectEntry::Vacant(no_entry)
+					} else {
+						DirectEntry::Occupied(BoolViewInner::Const(false))
+					}
+				}
 			},
-			EncodingType::Lazy => OrderStorage::Lazy(LazyOrderStorage {
-				min_index: 0,
-				max_index: 0,
-				lb_index: slv.engine_mut().state.trail.track_int(-1),
-				ub_index: slv.engine_mut().state.trail.track_int(-1),
-				storage: Vec::default(),
-			}),
-		};
-		let direct_encoding = match direct_encoding {
-			EncodingType::Eager => {
-				DirectStorage::Eager(slv.oracle.new_var_range(orig_domain_len - 2))
-			}
-			EncodingType::Lazy => DirectStorage::Lazy(HashMap::default()),
-		};
+		}
+	}
 
-		// Enforce consistency constraints for eager literals
-		if let OrderStorage::Eager { storage, .. } = &order_encoding {
-			let mut direct_enc_iter = if let DirectStorage::Eager(vars) = &direct_encoding {
-				Some(vars.clone())
-			} else {
-				None
-			}
-			.into_iter()
-			.flatten();
-			for (ord_i, ord_j) in storage.clone().tuple_windows() {
-				let ord_i: RawLit = ord_i.into(); // x<i
-				let ord_j: RawLit = ord_j.into(); // x<j, where j = i + n and n≥1
-				slv.oracle.add_clause([!ord_i, ord_j]).unwrap(); // x<i -> x<(i+n)
-				if matches!(direct_encoding, DirectStorage::Eager(_)) {
-					let eq_i: RawLit = direct_enc_iter.next().unwrap().into();
-					slv.oracle.add_clause([!eq_i, !ord_i]).unwrap(); // x=i -> x≥i
-					slv.oracle.add_clause([!eq_i, ord_j]).unwrap(); // x=i -> x<(i+n)
-					slv.oracle.add_clause([eq_i, ord_i, !ord_j]).unwrap(); // x≠i -> (x<i \/ x≥(i+n))
+	/// Return the [`BoolViewInner`] that represent the condition `= i`, if it
+	/// already exists.
+	///
+	/// The given `domain` is (in the case of eager creation) used to determine
+	/// the offset of the variable in the `VarRange`.
+	fn find(&self, domain: &RangeList<IntVal>, i: IntVal) -> Option<BoolViewInner> {
+		match self {
+			DirectStorage::Eager(vars) => {
+				// Calculate the offset in the VarRange
+				let mut offset = Some(-1); // -1 to account for the lower bound
+				for r in domain.iter() {
+					if i < *r.start() {
+						offset = None;
+						break;
+					} else if r.contains(&i) {
+						offset = Some(offset.unwrap() + i - r.start());
+						break;
+					} else {
+						offset = Some(offset.unwrap() + r.end() - r.start() + 1);
+					}
 				}
+				Some(if let Some(offset) = offset {
+					debug_assert!(
+						(offset as usize) < vars.len(),
+						"var range offset, {}, must be in [{}, {})",
+						offset,
+						0,
+						vars.len(),
+					);
+					BoolViewInner::Lit(vars.index(offset as usize).into())
+				} else {
+					BoolViewInner::Const(false)
+				})
 			}
-			debug_assert!(direct_enc_iter.next().is_none());
+			DirectStorage::Lazy(map) => map.get(&i).map(|v| BoolViewInner::Lit((*v).into())),
 		}
-
-		// Create the resulting integer variable
-		let iv = slv.engine_mut().state.int_vars.push(Self {
-			direct_encoding,
-			domain,
-			order_encoding,
-			upper_bound,
-		});
-		// Create propagator activation list
-		let r = slv
-			.engine_mut()
-			.state
-			.int_activation
-			.push(Default::default());
-		debug_assert_eq!(iv, r);
-
-		// Setup the boolean to integer mapping
-		if let OrderStorage::Eager { storage, .. } = &slv.engine().state.int_vars[iv].order_encoding
-		{
-			let mut vars = storage.clone();
-			if let DirectStorage::Eager(vars2) = &slv.engine().state.int_vars[iv].direct_encoding {
-				debug_assert_eq!(Into::<i32>::into(vars.end()) + 1, vars2.start().into());
-				vars = VarRange::new(vars.start(), vars2.end());
-			}
-			slv.engine_mut()
-				.state
-				.bool_to_int
-				.insert_eager(vars.clone(), iv);
-			slv.engine_mut()
-				.state
-				.trail
-				.grow_to_boolvar(vars.clone().last().unwrap());
-			for l in vars {
-				slv.oracle.add_observed_var(l);
-			}
-		}
-
-		IntView(IntViewInner::VarRef(iv))
 	}
+}
 
-	/// Returns the meaning of a literal in the context of this integer variable.
-	///
-	/// # Warning
-	///
-	/// This method can only be used with literals that were eagerly created for
-	/// this integer variable. Lazy literals should be mapped using
-	/// [`BoolToIntMap`].
-	pub(crate) fn lit_meaning(&self, lit: RawLit) -> LitMeaning {
-		let var = lit.var();
-		let ret = |l: LitMeaning| {
-			if lit.is_negated() {
-				!l
-			} else {
-				l
-			}
-		};
-
-		let OrderStorage::Eager { storage, .. } = &self.order_encoding else {
-			unreachable!("lit_meaning called on non-eager variable")
-		};
-		if storage.contains(&var) {
-			let mut offset = storage.find(var).unwrap() as IntVal + 1; // +1 because first value is not encoded
-			for r in self.domain.iter() {
-				let r_len = r.end() - r.start() + 1;
-				if offset < r_len {
-					return ret(LitMeaning::Less(*r.start() + offset));
-				}
-				offset -= r_len;
-			}
-			unreachable!()
-		}
-		let DirectStorage::Eager(vars) = &self.direct_encoding else {
-			unreachable!("lit_meaning called on non-eager variable")
-		};
-		debug_assert!(vars.contains(&var));
-		let mut offset = vars.find(var).unwrap() as IntVal + 1;
-		for r in self.domain.iter() {
-			let r_len = r.end() - r.start() + 1;
-			if offset < r_len {
-				return ret(LitMeaning::Eq(*r.start() + offset));
-			}
-			offset -= r_len;
-		}
-		unreachable!()
-	}
-
-	#[inline]
-	/// Method used to normalize a [`LitMeaning`] to find perform a lookup.
-	///
-	/// This method will replace any [`LitMeaning::Eq`] or [`LitMeaning::NotEq`]
-	/// meaning asking for the global lower or upper bound with the appropriate
-	/// [`LitMeaning::Less`] meaning. It will also negate any
-	/// [`LitMeaning::NotEq`] and [`LitMeaning::GreaterEq`] and return a boolean
-	/// indicating if the meaning was negated.
-	fn normalize_lit_meaning(
-		&self,
-		mut lit: LitMeaning,
-		lb: IntVal,
-		ub: IntVal,
-	) -> (LitMeaning, bool) {
-		let mut negate = false;
-		match lit {
-			LitMeaning::Eq(i) | LitMeaning::NotEq(i) if i == lb => {
-				negate = matches!(lit, LitMeaning::NotEq(_));
-				lit = LitMeaning::Less(lb + 1);
-			}
-			LitMeaning::Eq(i) | LitMeaning::NotEq(i) if i == ub => {
-				negate = matches!(lit, LitMeaning::Eq(_));
-				lit = LitMeaning::Less(ub);
-			}
-			LitMeaning::GreaterEq(_) | LitMeaning::NotEq(_) => {
-				lit = !lit;
-				negate = true;
-			}
-			_ => {}
-		}
-		(lit, negate)
-	}
-
+impl IntVar {
 	/// Access the Boolean literal with the given meaning, creating it if it is
 	/// not yet available.
 	pub(crate) fn bool_lit(
@@ -345,6 +404,22 @@ impl IntVar {
 			_ => unreachable!(),
 		});
 		if negate { !bv } else { bv }.into()
+	}
+
+	/// Returns the lower and upper bounds of the current state of the integer variable.
+	pub(crate) fn get_bounds(&self, trail: &impl TrailingActions) -> (IntVal, IntVal) {
+		let lb = match &self.order_encoding {
+			OrderStorage::Eager { lower_bound, .. } => trail.get_trailed_int(*lower_bound),
+			OrderStorage::Lazy(storage) => {
+				let low = trail.get_trailed_int(storage.lb_index);
+				if low >= 0 {
+					storage.storage[low as usize].val
+				} else {
+					*self.domain.lower_bound().unwrap()
+				}
+			}
+		};
+		(lb, trail.get_trailed_int(self.upper_bound))
 	}
 
 	/// Returns the boolean view associated with `≥ v` if it exists or weaker version otherwise.
@@ -507,20 +582,202 @@ impl IntVar {
 		}
 	}
 
-	/// Returns the lower and upper bounds of the current state of the integer variable.
-	pub(crate) fn get_bounds(&self, trail: &impl TrailingActions) -> (IntVal, IntVal) {
-		let lb = match &self.order_encoding {
-			OrderStorage::Eager { lower_bound, .. } => trail.get_trailed_int(*lower_bound),
-			OrderStorage::Lazy(storage) => {
-				let low = trail.get_trailed_int(storage.lb_index);
-				if low >= 0 {
-					storage.storage[low as usize].val
-				} else {
-					*self.domain.lower_bound().unwrap()
-				}
+	/// Returns the meaning of a literal in the context of this integer variable.
+	///
+	/// # Warning
+	///
+	/// This method can only be used with literals that were eagerly created for
+	/// this integer variable. Lazy literals should be mapped using
+	/// [`BoolToIntMap`].
+	pub(crate) fn lit_meaning(&self, lit: RawLit) -> LitMeaning {
+		let var = lit.var();
+		let ret = |l: LitMeaning| {
+			if lit.is_negated() {
+				!l
+			} else {
+				l
 			}
 		};
-		(lb, trail.get_trailed_int(self.upper_bound))
+
+		let OrderStorage::Eager { storage, .. } = &self.order_encoding else {
+			unreachable!("lit_meaning called on non-eager variable")
+		};
+		if storage.contains(&var) {
+			let mut offset = storage.find(var).unwrap() as IntVal + 1; // +1 because first value is not encoded
+			for r in self.domain.iter() {
+				let r_len = r.end() - r.start() + 1;
+				if offset < r_len {
+					return ret(LitMeaning::Less(*r.start() + offset));
+				}
+				offset -= r_len;
+			}
+			unreachable!()
+		}
+		let DirectStorage::Eager(vars) = &self.direct_encoding else {
+			unreachable!("lit_meaning called on non-eager variable")
+		};
+		debug_assert!(vars.contains(&var));
+		let mut offset = vars.find(var).unwrap() as IntVal + 1;
+		for r in self.domain.iter() {
+			let r_len = r.end() - r.start() + 1;
+			if offset < r_len {
+				return ret(LitMeaning::Eq(*r.start() + offset));
+			}
+			offset -= r_len;
+		}
+		unreachable!()
+	}
+	/// Create a new integer variable within the given solver, which the given
+	/// domain. The `order_encoding` and `direct_encoding` parameters determine
+	/// whether literals to reason about the integer variables are created eagerly
+	/// or lazily.
+	pub(crate) fn new_in<Oracle: PropagatingSolver<Engine>>(
+		slv: &mut Solver<Oracle>,
+		domain: IntSetVal,
+		order_encoding: EncodingType,
+		direct_encoding: EncodingType,
+	) -> IntView {
+		let orig_domain_len: usize = domain
+			.iter()
+			.map(|x| (x.end() - x.start() + 1) as usize)
+			.sum();
+		assert!(
+			orig_domain_len != 0,
+			"Unable to create integer variable empty domain"
+		);
+		if orig_domain_len == 1 {
+			return IntView(IntViewInner::Const(*domain.lower_bound().unwrap()));
+		}
+		let lb = *domain.lower_bound().unwrap();
+		let ub = *domain.upper_bound().unwrap();
+		if orig_domain_len == 2 {
+			let lit = slv.oracle.new_lit();
+			return IntView(IntViewInner::Bool {
+				transformer: LinearTransform {
+					scale: NonZeroIntVal::new(ub - lb).unwrap(),
+					offset: lb,
+				},
+				lit,
+			});
+		}
+		debug_assert!(
+			direct_encoding != EncodingType::Eager || order_encoding == EncodingType::Eager
+		);
+
+		let upper_bound = slv.engine_mut().state.trail.track_int(ub);
+		let order_encoding = match order_encoding {
+			EncodingType::Eager => OrderStorage::Eager {
+				lower_bound: slv.engine_mut().state.trail.track_int(lb),
+				storage: slv.oracle.new_var_range(orig_domain_len - 1),
+			},
+			EncodingType::Lazy => OrderStorage::Lazy(LazyOrderStorage {
+				min_index: 0,
+				max_index: 0,
+				lb_index: slv.engine_mut().state.trail.track_int(-1),
+				ub_index: slv.engine_mut().state.trail.track_int(-1),
+				storage: Vec::default(),
+			}),
+		};
+		let direct_encoding = match direct_encoding {
+			EncodingType::Eager => {
+				DirectStorage::Eager(slv.oracle.new_var_range(orig_domain_len - 2))
+			}
+			EncodingType::Lazy => DirectStorage::Lazy(HashMap::default()),
+		};
+
+		// Enforce consistency constraints for eager literals
+		if let OrderStorage::Eager { storage, .. } = &order_encoding {
+			let mut direct_enc_iter = if let DirectStorage::Eager(vars) = &direct_encoding {
+				Some(vars.clone())
+			} else {
+				None
+			}
+			.into_iter()
+			.flatten();
+			for (ord_i, ord_j) in storage.clone().tuple_windows() {
+				let ord_i: RawLit = ord_i.into(); // x<i
+				let ord_j: RawLit = ord_j.into(); // x<j, where j = i + n and n≥1
+				slv.oracle.add_clause([!ord_i, ord_j]).unwrap(); // x<i -> x<(i+n)
+				if matches!(direct_encoding, DirectStorage::Eager(_)) {
+					let eq_i: RawLit = direct_enc_iter.next().unwrap().into();
+					slv.oracle.add_clause([!eq_i, !ord_i]).unwrap(); // x=i -> x≥i
+					slv.oracle.add_clause([!eq_i, ord_j]).unwrap(); // x=i -> x<(i+n)
+					slv.oracle.add_clause([eq_i, ord_i, !ord_j]).unwrap(); // x≠i -> (x<i \/ x≥(i+n))
+				}
+			}
+			debug_assert!(direct_enc_iter.next().is_none());
+		}
+
+		// Create the resulting integer variable
+		let iv = slv.engine_mut().state.int_vars.push(Self {
+			direct_encoding,
+			domain,
+			order_encoding,
+			upper_bound,
+		});
+		// Create propagator activation list
+		let r = slv
+			.engine_mut()
+			.state
+			.int_activation
+			.push(Default::default());
+		debug_assert_eq!(iv, r);
+
+		// Setup the boolean to integer mapping
+		if let OrderStorage::Eager { storage, .. } = &slv.engine().state.int_vars[iv].order_encoding
+		{
+			let mut vars = storage.clone();
+			if let DirectStorage::Eager(vars2) = &slv.engine().state.int_vars[iv].direct_encoding {
+				debug_assert_eq!(Into::<i32>::into(vars.end()) + 1, vars2.start().into());
+				vars = VarRange::new(vars.start(), vars2.end());
+			}
+			slv.engine_mut()
+				.state
+				.bool_to_int
+				.insert_eager(vars.clone(), iv);
+			slv.engine_mut()
+				.state
+				.trail
+				.grow_to_boolvar(vars.clone().last().unwrap());
+			for l in vars {
+				slv.oracle.add_observed_var(l);
+			}
+		}
+
+		IntView(IntViewInner::VarRef(iv))
+	}
+
+	#[inline]
+	/// Method used to normalize a [`LitMeaning`] to find perform a lookup.
+	///
+	/// This method will replace any [`LitMeaning::Eq`] or [`LitMeaning::NotEq`]
+	/// meaning asking for the global lower or upper bound with the appropriate
+	/// [`LitMeaning::Less`] meaning. It will also negate any
+	/// [`LitMeaning::NotEq`] and [`LitMeaning::GreaterEq`] and return a boolean
+	/// indicating if the meaning was negated.
+	fn normalize_lit_meaning(
+		&self,
+		mut lit: LitMeaning,
+		lb: IntVal,
+		ub: IntVal,
+	) -> (LitMeaning, bool) {
+		let mut negate = false;
+		match lit {
+			LitMeaning::Eq(i) | LitMeaning::NotEq(i) if i == lb => {
+				negate = matches!(lit, LitMeaning::NotEq(_));
+				lit = LitMeaning::Less(lb + 1);
+			}
+			LitMeaning::Eq(i) | LitMeaning::NotEq(i) if i == ub => {
+				negate = matches!(lit, LitMeaning::Eq(_));
+				lit = LitMeaning::Less(ub);
+			}
+			LitMeaning::GreaterEq(_) | LitMeaning::NotEq(_) => {
+				lit = !lit;
+				negate = true;
+			}
+			_ => {}
+		}
+		(lit, negate)
 	}
 
 	/// Notify that a new lower bound has been propagated for the variable,
@@ -622,128 +879,6 @@ impl IntVar {
 	}
 }
 
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[allow(
-	variant_size_differences,
-	reason = "TODO: Investigate if using Box improves performance"
-)]
-/// The storage used to store the variables for the inequality conditions.
-pub(crate) enum OrderStorage {
-	/// Variables for all inequality conditions are eagerly created and stored in
-	/// order.
-	Eager {
-		/// A trailed integer that represents the currently lower bound of the
-		/// variable.
-		lower_bound: TrailedInt,
-		/// The range of Boolean variables that represent the inequality conditions.
-		storage: VarRange,
-	},
-	/// Variables for inequality conditions are lazily created and specialized
-	/// noded structure, a [`LazyOrderStorage`].
-	Lazy(LazyOrderStorage),
-}
-
-/// The definition given to a lazily created literal.
-pub(crate) struct LazyLitDef {
-	/// The meaning that the literal is meant to represent.
-	pub(crate) meaning: LitMeaning,
-	/// The variable that represent:
-	/// - if `meaning` is `LitMeaning::Less(j)`, then `prev` contains the literal
-	///   `< i` where `i` is the value right before `j` in the storage.
-	/// - if `meaning` is `LitMeaning::Eq(k)`, then `prev` contains the literal
-	///   `<j`.
-	pub(crate) prev: Option<RawVar>,
-	/// The variable that represent the literal `< k` where `k` is the value right
-	/// after the value represented by the literal.
-	pub(crate) next: Option<RawVar>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-/// A storage structure to manage lazily created order literals for an integer
-/// variable.
-pub(crate) struct LazyOrderStorage {
-	/// The index of the node with the minimum value in the storage.
-	min_index: u32,
-	/// The index of the node with the maximum value in the storage.
-	max_index: u32,
-	/// The index of the node that currently represents the lower bound of the
-	/// integer variable.
-	lb_index: TrailedInt,
-	/// The index of the node that currently represents the upper bound of the
-	/// integer variable.
-	ub_index: TrailedInt,
-	/// The storage of all currently created nodes containing the order literals
-	/// for the integer variable.
-	storage: Vec<OrderNode>,
-}
-
-/// Type alias for an iterator that yields the ranges of a [`RangeList`], which
-/// is used to represent the domains of an integer variable.
-type RangeIter<'a> = Peekable<
-	Map<
-		<&'a RangeList<IntVal> as IntoIterator>::IntoIter,
-		fn(RangeInclusive<&'a IntVal>) -> RangeInclusive<IntVal>,
-	>,
->;
-
-#[derive(Debug)]
-/// An entry in [`OrderStorage`] that can be used to access the representation
-/// of an inequality condition, or insert a new literal to represent the
-/// condition otherwise.
-enum OrderEntry<'a> {
-	/// Entry already exists and was eagerly created.
-	Eager(&'a VarRange, usize),
-	/// Entry already exists and was lazily created.
-	Occupied {
-		/// Reference to the storage where the entry is stored.
-		storage: &'a mut LazyOrderStorage,
-		/// The index of the node in the storage that the entry points to.
-		index: u32,
-		/// An iterator pointing at the range in the domain in which the value of
-		/// which the value of the entry is part.
-		range_iter: RangeIter<'a>,
-	},
-	/// Entry does not exist and can be lazily created.
-	Vacant {
-		/// Reference to the storage where the new entry will be created.
-		storage: &'a mut LazyOrderStorage,
-		/// The index of the node that contains the value right before the new entry
-		/// that will be created.
-		prev_index: IntVal,
-		/// An iterator pointing at the range in the domain in which the value of
-		/// which the value of the new entry is part.
-		range_iter: RangeIter<'a>,
-		/// The value for which the entry will be created.
-		val: IntVal,
-	},
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-/// Type useed to store individual entries in [`LazyOrderStorage`].
-pub(crate) struct OrderNode {
-	/// The value for which `var` represents `x < val`.
-	val: IntVal,
-	/// The variable representing `x < val`.
-	var: RawVar,
-	/// Whether there is a node with a value less than `val`.
-	has_prev: bool,
-	/// The index of the node with a value less than `val`.
-	prev: u32,
-	/// Whether there is a node with a value greater than `val`.
-	has_next: bool,
-	/// The index of the node with a value greater than `val`.
-	next: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-/// A direction to search in.
-enum SearchDirection {
-	/// Search from low to high.
-	Increasing,
-	/// Search from high to low.
-	Decreasing,
-}
-
 impl LazyOrderStorage {
 	/// Find the the index of the node that contains the value or the node
 	/// "before" the value.
@@ -769,16 +904,6 @@ impl LazyOrderStorage {
 		self.storage.is_empty()
 	}
 
-	/// Returns the node with the minimum [`OrderNode::val`] present in the
-	/// storage, or [`None`] if the storage is empty.
-	fn min(&self) -> Option<&OrderNode> {
-		if self.is_empty() {
-			None
-		} else {
-			Some(&self[self.min_index])
-		}
-	}
-
 	/// Returns the node with the maximum [`OrderNode::val`] present in the
 	/// storage, or [`None`] if the storage is empty.
 	fn max(&self) -> Option<&OrderNode> {
@@ -788,10 +913,21 @@ impl LazyOrderStorage {
 			Some(&self[self.max_index])
 		}
 	}
+
+	/// Returns the node with the minimum [`OrderNode::val`] present in the
+	/// storage, or [`None`] if the storage is empty.
+	fn min(&self) -> Option<&OrderNode> {
+		if self.is_empty() {
+			None
+		} else {
+			Some(&self[self.min_index])
+		}
+	}
 }
 
 impl Index<u32> for LazyOrderStorage {
 	type Output = OrderNode;
+
 	fn index(&self, index: u32) -> &Self::Output {
 		&self.storage[index as usize]
 	}
@@ -803,112 +939,132 @@ impl IndexMut<u32> for LazyOrderStorage {
 	}
 }
 
-impl OrderStorage {
-	/// Returns the lowest integer value `j`, for which `< i` is equivalent to `<
-	/// j` in the given `domain. In addition it returns the index of the range in
-	/// `domain` in which `j` is located, and calculate the offset of the
-	/// representation `< j` in a VarRange when the order literals are eagerly
-	/// created.
-	fn resolve_val(domain: &RangeList<IntVal>, val: IntVal) -> (IntVal, usize, RangeIter) {
-		let mut offset = -1; // -1 to account for the lower bound
-		let mut it = domain.iter().peekable();
-		let mut real_val = val;
-		loop {
-			let r = it.peek().unwrap();
-			if val < *r.start() {
-				real_val = *r.start();
-				break;
-			} else if r.contains(&val) {
-				offset += val - r.start();
-				break;
-			} else {
-				offset += r.end() - r.start() + 1;
-			}
-			let _ = it.next().unwrap();
-		}
-		(real_val, offset as usize, it)
-	}
-
-	/// Locate the position in the [`OrderStorage`] that would be used to store
-	/// the representation of the condition `< i`. The method will return a
-	/// [`OrderEntry`] object that can be used to access the condition as a
-	/// [`RawVar`] if it already exists, or insert a new literal to represent the
-	/// condition otherwise.
+impl LitMeaning {
+	/// Returns the clauses that can be used to define the given literal according
+	/// to the meaning `self`.
 	///
-	/// The given `domain` is (in the case of eager creation) used to determine
-	/// the offset of the variable in the `VarRange`.
-	fn entry<'a>(&'a mut self, domain: &'a RangeList<IntVal>, val: IntVal) -> OrderEntry<'a> {
-		let (val, offset, range_iter) = Self::resolve_val(domain, val);
-
+	/// Note this method is only intended to be used to define positive literals,
+	/// and it is thus assumed to be unreachable to be called on
+	/// [`LitMeaning::NotEq`] or [`LitMeaning::GreaterEq`].
+	pub(crate) fn defining_clauses(
+		&self,
+		lit: RawLit,
+		prev: Option<RawLit>,
+		next: Option<RawLit>,
+	) -> Vec<Clause> {
+		let mut ret = Vec::<Clause>::new();
 		match self {
-			OrderStorage::Eager { storage, .. } => OrderEntry::Eager(storage, offset),
-			OrderStorage::Lazy(storage) => {
-				if storage.is_empty() || storage.min().unwrap().val > val {
-					return OrderEntry::Vacant {
-						storage,
-						prev_index: -1,
-						range_iter,
-						val,
-					};
-				} else if storage.max().unwrap().val < val {
-					return OrderEntry::Vacant {
-						prev_index: storage.max_index as IntVal,
-						storage,
-						range_iter,
-						val,
-					};
-				}
+			LitMeaning::Eq(_) => {
+				let prev = prev.expect("prev should contain the GreaterEq literal for the value"); // x≥i
+				let next =
+					next.expect("next should contain the GreaterEq literal for the next value"); // x≥i+n
 
-				let i = storage.find_index(storage.min_index, SearchDirection::Increasing, val);
-				debug_assert!(storage[i].val <= val);
-				if storage[i].val == val {
-					OrderEntry::Occupied {
-						storage,
-						index: i,
-						range_iter,
-					}
-				} else {
-					OrderEntry::Vacant {
-						storage,
-						prev_index: i as IntVal,
-						range_iter,
-						val,
-					}
+				ret.push(vec![!lit, !prev]); // x=i -> x≥i
+				ret.push(vec![!lit, next]); // x=i -> x<i+n
+				ret.push(vec![lit, prev, !next]); // x!=i -> x<i \/ x>i+n
+			}
+			LitMeaning::Less(_) => {
+				if let Some(prev) = prev {
+					ret.push(vec![!prev, lit]); // x<(i-n) -> x<i
+				}
+				if let Some(next) = next {
+					ret.push(vec![!lit, next]); // x<i -> x<(i+n)
 				}
 			}
+			_ => unreachable!(),
 		}
+		ret
 	}
+}
 
-	/// Return the [`RawVar`] that represent the condition `< i`, if it already
-	/// exists.
-	///
-	/// The given `domain` is (in the case of eager creation) used to determine
-	/// the offset of the variable in the `VarRange`.
-	fn find(&self, domain: &RangeList<IntVal>, val: IntVal) -> Option<RawVar> {
-		let (val, offset, _) = Self::resolve_val(domain, val);
+impl Not for LitMeaning {
+	type Output = LitMeaning;
 
+	fn not(self) -> Self::Output {
 		match self {
-			OrderStorage::Eager { storage, .. } => Some(storage.index(offset)),
-			OrderStorage::Lazy(storage) => {
-				if storage.is_empty()
-					|| storage.min().unwrap().val > val
-					|| storage.max().unwrap().val < val
-				{
-					return None;
-				}
-
-				let i = storage.find_index(storage.min_index, SearchDirection::Increasing, val);
-				if storage[i].val == val {
-					Some(storage[i].var)
-				} else {
-					None
-				}
-			}
+			LitMeaning::Eq(i) => LitMeaning::NotEq(i),
+			LitMeaning::NotEq(i) => LitMeaning::Eq(i),
+			LitMeaning::GreaterEq(i) => LitMeaning::Less(i),
+			LitMeaning::Less(i) => LitMeaning::GreaterEq(i),
 		}
 	}
 }
 
 impl OrderEntry<'_> {
+	/// Forward the entry to the entry for next value in the domain.
+	///
+	/// Note that it is assumed that a next value exists in the domain, and this
+	/// method will panic otherwise.
+	fn next_value(self) -> Self {
+		match self {
+			OrderEntry::Eager(vars, offset) => OrderEntry::Eager(vars, offset + 1),
+			OrderEntry::Occupied {
+				storage,
+				index,
+				mut range_iter,
+			} => {
+				let next = storage[index].val + 1;
+				let next = if range_iter.peek().unwrap().contains(&next) {
+					next
+				} else {
+					let _ = range_iter.next().unwrap();
+					*range_iter.peek().unwrap().start()
+				};
+				let next_index = storage[index].next;
+				if storage[index].has_next && storage[next_index].val == next {
+					OrderEntry::Occupied {
+						storage,
+						index: next_index,
+						range_iter,
+					}
+				} else {
+					OrderEntry::Vacant {
+						storage,
+						prev_index: index as IntVal,
+						range_iter,
+						val: next,
+					}
+				}
+			}
+			OrderEntry::Vacant {
+				storage,
+				prev_index,
+				mut range_iter,
+				val,
+			} => {
+				let next = val + 1;
+				let next = if range_iter.peek().unwrap().contains(&next) {
+					next
+				} else {
+					let _ = range_iter.next().unwrap();
+					*range_iter.peek().unwrap().start()
+				};
+				if prev_index >= 0
+					&& storage[prev_index as u32].has_next
+					&& storage[storage[prev_index as u32].next].val == next
+				{
+					OrderEntry::Occupied {
+						index: storage[prev_index as u32].next,
+						storage,
+						range_iter,
+					}
+				} else if !storage.is_empty() && storage.min().unwrap().val == next {
+					OrderEntry::Occupied {
+						index: storage.min_index,
+						storage,
+						range_iter,
+					}
+				} else {
+					OrderEntry::Vacant {
+						storage,
+						prev_index,
+						range_iter,
+						val: next,
+					}
+				}
+			}
+		}
+	}
 	/// Extract the [`RawVar`] if the entry is occupied, or insert a new
 	/// variable using the given function.
 	///
@@ -1009,273 +1165,115 @@ impl OrderEntry<'_> {
 			}
 		}
 	}
+}
 
-	/// Forward the entry to the entry for next value in the domain.
+impl OrderStorage {
+	/// Locate the position in the [`OrderStorage`] that would be used to store
+	/// the representation of the condition `< i`. The method will return a
+	/// [`OrderEntry`] object that can be used to access the condition as a
+	/// [`RawVar`] if it already exists, or insert a new literal to represent the
+	/// condition otherwise.
 	///
-	/// Note that it is assumed that a next value exists in the domain, and this
-	/// method will panic otherwise.
-	fn next_value(self) -> Self {
+	/// The given `domain` is (in the case of eager creation) used to determine
+	/// the offset of the variable in the `VarRange`.
+	fn entry<'a>(&'a mut self, domain: &'a RangeList<IntVal>, val: IntVal) -> OrderEntry<'a> {
+		let (val, offset, range_iter) = Self::resolve_val(domain, val);
+
 		match self {
-			OrderEntry::Eager(vars, offset) => OrderEntry::Eager(vars, offset + 1),
-			OrderEntry::Occupied {
-				storage,
-				index,
-				mut range_iter,
-			} => {
-				let next = storage[index].val + 1;
-				let next = if range_iter.peek().unwrap().contains(&next) {
-					next
-				} else {
-					let _ = range_iter.next().unwrap();
-					*range_iter.peek().unwrap().start()
-				};
-				let next_index = storage[index].next;
-				if storage[index].has_next && storage[next_index].val == next {
+			OrderStorage::Eager { storage, .. } => OrderEntry::Eager(storage, offset),
+			OrderStorage::Lazy(storage) => {
+				if storage.is_empty() || storage.min().unwrap().val > val {
+					return OrderEntry::Vacant {
+						storage,
+						prev_index: -1,
+						range_iter,
+						val,
+					};
+				} else if storage.max().unwrap().val < val {
+					return OrderEntry::Vacant {
+						prev_index: storage.max_index as IntVal,
+						storage,
+						range_iter,
+						val,
+					};
+				}
+
+				let i = storage.find_index(storage.min_index, SearchDirection::Increasing, val);
+				debug_assert!(storage[i].val <= val);
+				if storage[i].val == val {
 					OrderEntry::Occupied {
 						storage,
-						index: next_index,
+						index: i,
 						range_iter,
 					}
 				} else {
 					OrderEntry::Vacant {
 						storage,
-						prev_index: index as IntVal,
+						prev_index: i as IntVal,
 						range_iter,
-						val: next,
+						val,
 					}
 				}
 			}
-			OrderEntry::Vacant {
-				storage,
-				prev_index,
-				mut range_iter,
-				val,
-			} => {
-				let next = val + 1;
-				let next = if range_iter.peek().unwrap().contains(&next) {
-					next
-				} else {
-					let _ = range_iter.next().unwrap();
-					*range_iter.peek().unwrap().start()
-				};
-				if prev_index >= 0
-					&& storage[prev_index as u32].has_next
-					&& storage[storage[prev_index as u32].next].val == next
+		}
+	}
+
+	/// Return the [`RawVar`] that represent the condition `< i`, if it already
+	/// exists.
+	///
+	/// The given `domain` is (in the case of eager creation) used to determine
+	/// the offset of the variable in the `VarRange`.
+	fn find(&self, domain: &RangeList<IntVal>, val: IntVal) -> Option<RawVar> {
+		let (val, offset, _) = Self::resolve_val(domain, val);
+
+		match self {
+			OrderStorage::Eager { storage, .. } => Some(storage.index(offset)),
+			OrderStorage::Lazy(storage) => {
+				if storage.is_empty()
+					|| storage.min().unwrap().val > val
+					|| storage.max().unwrap().val < val
 				{
-					OrderEntry::Occupied {
-						index: storage[prev_index as u32].next,
-						storage,
-						range_iter,
-					}
-				} else if !storage.is_empty() && storage.min().unwrap().val == next {
-					OrderEntry::Occupied {
-						index: storage.min_index,
-						storage,
-						range_iter,
-					}
+					return None;
+				}
+
+				let i = storage.find_index(storage.min_index, SearchDirection::Increasing, val);
+				if storage[i].val == val {
+					Some(storage[i].var)
 				} else {
-					OrderEntry::Vacant {
-						storage,
-						prev_index,
-						range_iter,
-						val: next,
-					}
+					None
 				}
 			}
 		}
 	}
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-/// The structure that stores the equality conditions. Equality conditions can
-/// either be eagerly crated, and stored as a range of variables, or lazily
-/// created and stored in a [`HashMap`] once created.
-pub(crate) enum DirectStorage {
-	/// Variables for all equality conditions are eagerly created and stored in order
-	Eager(VarRange),
-	/// Variables for equality conditions are lazily created and stored in a hashmap
-	Lazy(HashMap<IntVal, RawVar>),
-}
-
-/// An entry in the [`DirectStorage`] that can be used to access the
-/// representation of an equality condition, or insert a new literal to
-/// represent the condition otherwise.
-enum DirectEntry<'a> {
-	/// The condition is already stored in the [`DirectStorage`].
-	Occupied(BoolViewInner),
-	/// The condition is not yet stored in the [`DirectStorage`].
-	Vacant(VacantEntry<'a, IntVal, RawVar>),
-}
-
-impl DirectStorage {
-	/// Locate the position in the [`DirectStorage`] that would be used to store
-	/// the representation of the condition `= i`. The method will return a
-	/// [`DirectEntry`] object that can be used to access the condition as a
-	/// [`BoolViewInner`] if it already exists, or insert a new literal to
-	/// represent the condition otherwise.
-	///
-	/// The given `domain` is (in the case of eager creation) used to determine
-	/// the offset of the variable in the `VarRange`.
-	fn entry(&mut self, domain: &RangeList<IntVal>, i: IntVal) -> DirectEntry<'_> {
-		match self {
-			DirectStorage::Eager(vars) => {
-				// Calculate the offset in the VarRange
-				let mut offset = Some(-1); // -1 to account for the lower bound
-				for r in domain.iter() {
-					if i < *r.start() {
-						offset = None;
-						break;
-					} else if r.contains(&i) {
-						offset = Some(offset.unwrap() + i - r.start());
-						break;
-					} else {
-						offset = Some(offset.unwrap() + r.end() - r.start() + 1);
-					}
-				}
-				if let Some(offset) = offset {
-					debug_assert!(
-						(offset as usize) < vars.len(),
-						"var range offset, {}, must be in [{}, {})",
-						offset,
-						0,
-						vars.len(),
-					);
-					DirectEntry::Occupied(BoolViewInner::Lit(vars.index(offset as usize).into()))
-				} else {
-					DirectEntry::Occupied(BoolViewInner::Const(false))
-				}
+	/// Returns the lowest integer value `j`, for which `< i` is equivalent to `<
+	/// j` in the given `domain. In addition it returns the index of the range in
+	/// `domain` in which `j` is located, and calculate the offset of the
+	/// representation `< j` in a VarRange when the order literals are eagerly
+	/// created.
+	fn resolve_val(domain: &RangeList<IntVal>, val: IntVal) -> (IntVal, usize, RangeIter) {
+		let mut offset = -1; // -1 to account for the lower bound
+		let mut it = domain.iter().peekable();
+		let mut real_val = val;
+		loop {
+			let r = it.peek().unwrap();
+			if val < *r.start() {
+				real_val = *r.start();
+				break;
+			} else if r.contains(&val) {
+				offset += val - r.start();
+				break;
+			} else {
+				offset += r.end() - r.start() + 1;
 			}
-			DirectStorage::Lazy(map) => match map.entry(i) {
-				hash_map::Entry::Occupied(entry) => {
-					DirectEntry::Occupied(BoolViewInner::Lit((*entry.get()).into()))
-				}
-				hash_map::Entry::Vacant(no_entry) => {
-					if domain.contains(&i) {
-						DirectEntry::Vacant(no_entry)
-					} else {
-						DirectEntry::Occupied(BoolViewInner::Const(false))
-					}
-				}
-			},
+			let _ = it.next().unwrap();
 		}
-	}
-
-	/// Return the [`BoolViewInner`] that represent the condition `= i`, if it
-	/// already exists.
-	///
-	/// The given `domain` is (in the case of eager creation) used to determine
-	/// the offset of the variable in the `VarRange`.
-	fn find(&self, domain: &RangeList<IntVal>, i: IntVal) -> Option<BoolViewInner> {
-		match self {
-			DirectStorage::Eager(vars) => {
-				// Calculate the offset in the VarRange
-				let mut offset = Some(-1); // -1 to account for the lower bound
-				for r in domain.iter() {
-					if i < *r.start() {
-						offset = None;
-						break;
-					} else if r.contains(&i) {
-						offset = Some(offset.unwrap() + i - r.start());
-						break;
-					} else {
-						offset = Some(offset.unwrap() + r.end() - r.start() + 1);
-					}
-				}
-				Some(if let Some(offset) = offset {
-					debug_assert!(
-						(offset as usize) < vars.len(),
-						"var range offset, {}, must be in [{}, {})",
-						offset,
-						0,
-						vars.len(),
-					);
-					BoolViewInner::Lit(vars.index(offset as usize).into())
-				} else {
-					BoolViewInner::Const(false)
-				})
-			}
-			DirectStorage::Lazy(map) => map.get(&i).map(|v| BoolViewInner::Lit((*v).into())),
-		}
+		(real_val, offset as usize, it)
 	}
 }
 
-impl DirectEntry<'_> {
-	/// Extract the [`BoolViewInner`] if the entry is occupied, or insert a new
-	/// variable using the given function.
-	fn or_insert_with(self, f: impl FnOnce() -> RawVar) -> BoolViewInner {
-		match self {
-			DirectEntry::Occupied(bv) => bv,
-			DirectEntry::Vacant(no_entry) => {
-				let v = f();
-				let _ = no_entry.insert(v);
-				BoolViewInner::Lit(v.into())
-			}
-		}
-	}
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-/// The meaning of a literal in the context of a [`IntVar`] `x`.
-pub enum LitMeaning {
-	/// Literal representing the condition `x = i`.
-	Eq(IntVal),
-	/// Literal representing the condition `x ≠ i`.
-	NotEq(IntVal),
-	/// Literal representing the condition `x ≥ i`.
-	GreaterEq(IntVal),
-	/// Literal representing the condition `x < i`.
-	Less(IntVal),
-}
-
-impl LitMeaning {
-	/// Returns the clauses that can be used to define the given literal according
-	/// to the meaning `self`.
-	///
-	/// Note this method is only intended to be used to define positive literals,
-	/// and it is thus assumed to be unreachable to be called on
-	/// [`LitMeaning::NotEq`] or [`LitMeaning::GreaterEq`].
-	pub(crate) fn defining_clauses(
-		&self,
-		lit: RawLit,
-		prev: Option<RawLit>,
-		next: Option<RawLit>,
-	) -> Vec<Clause> {
-		let mut ret = Vec::<Clause>::new();
-		match self {
-			LitMeaning::Eq(_) => {
-				let prev = prev.expect("prev should contain the GreaterEq literal for the value"); // x≥i
-				let next =
-					next.expect("next should contain the GreaterEq literal for the next value"); // x≥i+n
-
-				ret.push(vec![!lit, !prev]); // x=i -> x≥i
-				ret.push(vec![!lit, next]); // x=i -> x<i+n
-				ret.push(vec![lit, prev, !next]); // x!=i -> x<i \/ x>i+n
-			}
-			LitMeaning::Less(_) => {
-				if let Some(prev) = prev {
-					ret.push(vec![!prev, lit]); // x<(i-n) -> x<i
-				}
-				if let Some(next) = next {
-					ret.push(vec![!lit, next]); // x<i -> x<(i+n)
-				}
-			}
-			_ => unreachable!(),
-		}
-		ret
-	}
-}
-
-impl Not for LitMeaning {
-	type Output = LitMeaning;
-
-	fn not(self) -> Self::Output {
-		match self {
-			LitMeaning::Eq(i) => LitMeaning::NotEq(i),
-			LitMeaning::NotEq(i) => LitMeaning::Eq(i),
-			LitMeaning::GreaterEq(i) => LitMeaning::Less(i),
-			LitMeaning::Less(i) => LitMeaning::GreaterEq(i),
-		}
-	}
+index_vec::define_index_type! {
+	/// Identifies an integer variable in a [`Solver`]
+	pub struct IntVarRef = u32;
 }
 
 #[cfg(test)]
