+++ crates/huub/src/solver/poster.rs
@@ -9,12 +9,25 @@ use crate::{
 	ReformulationError,
 };
 
-/// Type alias to represent [`Propagator`] contained in a [`Box`], that is used
-/// by [`Engine`].
-pub(crate) type BoxedPropagator = Box<dyn for<'a> Propagator<SolvingContext<'a>, State>>;
 /// Type alias to represent [`Brancher`] contained in a [`Box`], that is used by
 /// [`Engine`].
 pub(crate) type BoxedBrancher = Box<dyn for<'a> Brancher<SolvingContext<'a>>>;
+
+/// Type alias to represent [`Propagator`] contained in a [`Box`], that is used
+/// by [`Engine`].
+pub(crate) type BoxedPropagator = Box<dyn for<'a> Propagator<SolvingContext<'a>, State>>;
+
+/// The trait used to register a brancher with the solver.
+pub(crate) trait BrancherPoster {
+	/// Register the brancher with the solver.
+	///
+	/// This method is expected to return the brancher to be registered.
+	///
+	/// The post method is given access to the solver's initialization actions,
+	/// which includes the ability to subscribe to variable events, creating
+	/// trailed data structures, and inspecting the current state of varaibles.
+	fn post<I: InitializationActions>(self, actions: &mut I) -> BoxedBrancher;
+}
 
 /// The trait used called to registering a propagator with the solver.
 pub(crate) trait Poster {
@@ -40,16 +53,4 @@ pub(crate) struct QueuePreferences {
 	pub(crate) enqueue_on_post: bool,
 	/// Priority level in the queue used for the propagator
 	pub(crate) priority: PriorityLevel,
-}
-
-/// The trait used to register a brancher with the solver.
-pub(crate) trait BrancherPoster {
-	/// Register the brancher with the solver.
-	///
-	/// This method is expected to return the brancher to be registered.
-	///
-	/// The post method is given access to the solver's initialization actions,
-	/// which includes the ability to subscribe to variable events, creating
-	/// trailed data structures, and inspecting the current state of varaibles.
-	fn post<I: InitializationActions>(self, actions: &mut I) -> BoxedBrancher;
 }
